# Conflicts:
#	src/main/java/bio/terra/app/controller/DatasetsApiController.java
#	src/main/java/bio/terra/service/dataset/flight/create/CreateDatasetAssetStep.java
#	src/main/java/bio/terra/service/dataset/flight/datadelete/DataDeletionStep.java
#	src/main/java/bio/terra/service/dataset/flight/delete/DeleteDatasetAuthzResource.java
#	src/main/java/bio/terra/service/filedata/flight/delete/DeleteFileAzureDirectoryStep.java
#	src/main/java/bio/terra/service/snapshot/flight/create/CreateSnapshotMetadataStep.java
#	src/main/java/bio/terra/service/snapshot/flight/create/CreateSnapshotPrimaryDataAssetGcpStep.java
#	src/main/java/bio/terra/service/snapshot/flight/create/CreateSnapshotPrimaryDataRowIdsStep.java
#	src/main/java/bio/terra/service/snapshot/flight/delete/DeleteSnapshotMetadataStep.java

+++ src/test/java/bio/terra/service/resourcemanagement/google/BucketResourceTest.java
@@ -1,7 +1,10 @@
 package bio.terra.service.resourcemanagement.google;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.equalToIgnoringCase;
+import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.matchesPattern;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -10,6 +13,8 @@ import static org.junit.Assert.assertTrue;
 import bio.terra.app.configuration.ConnectedTestConfiguration;
 import bio.terra.app.model.GoogleRegion;
 import bio.terra.buffer.model.ResourceInfo;
+import bio.terra.common.CollectionType;
+import bio.terra.common.EmbeddedDatabaseTest;
 import bio.terra.common.category.Connected;
 import bio.terra.common.fixtures.ConnectedOperations;
 import bio.terra.common.fixtures.JsonLoader;
@@ -17,10 +22,11 @@ import bio.terra.common.fixtures.Names;
 import bio.terra.model.BillingProfileModel;
 import bio.terra.model.DatasetRequestModel;
 import bio.terra.model.DatasetSummaryModel;
+import bio.terra.service.auth.iam.IamProviderInterface;
+import bio.terra.service.auth.iam.IamResourceType;
 import bio.terra.service.configuration.ConfigEnum;
 import bio.terra.service.configuration.ConfigurationService;
 import bio.terra.service.dataset.DatasetBucketDao;
-import bio.terra.service.iam.IamProviderInterface;
 import bio.terra.service.load.LoadDao;
 import bio.terra.service.profile.ProfileService;
 import bio.terra.service.resourcemanagement.BucketResourceLockTester;
@@ -28,14 +34,19 @@ import bio.terra.service.resourcemanagement.BufferService;
 import bio.terra.service.resourcemanagement.exception.GoogleResourceNotFoundException;
 import bio.terra.service.snapshot.exception.CorruptMetadataException;
 import com.google.api.client.util.Lists;
+import com.google.cloud.Binding;
+import com.google.cloud.Policy;
 import com.google.cloud.storage.Bucket;
+import com.google.cloud.storage.BucketInfo;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageOptions;
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Before;
@@ -56,6 +67,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 @ActiveProfiles({"google", "connectedtest"})
 @Category(Connected.class)
+@EmbeddedDatabaseTest
 public class BucketResourceTest {
   private final Logger logger = LoggerFactory.getLogger(BucketResourceTest.class);
 
@@ -77,7 +89,6 @@ public class BucketResourceTest {
   private BillingProfileModel profile;
   private Storage storage;
   private List<GoogleBucketResource> bucketResources;
-  private boolean allowReuseExistingBuckets;
   private GoogleProjectResource projectResource;
   private UUID datasetId;
 
@@ -101,7 +112,8 @@ public class BucketResourceTest {
         jsonLoader.loadObject(resourcePath, DatasetRequestModel.class);
     datasetRequest
         .name(Names.randomizeName(datasetRequest.getName()))
-        .defaultProfileId(profile.getId());
+        .defaultProfileId(profile.getId())
+        .dedicatedIngestServiceAccount(false);
 
     DatasetSummaryModel summaryModel = connectedOperations.createDataset(datasetRequest);
     datasetId = summaryModel.getId();
@@ -109,9 +121,7 @@ public class BucketResourceTest {
 
   @After
   public void teardown() throws Exception {
-    for (GoogleBucketResource bucketResource : bucketResources) {
-      deleteBucket(bucketResource);
-    }
+    bucketResources.forEach(this::deleteBucket);
     // Connected operations resets the configuration
     connectedOperations.teardown();
   }
@@ -119,12 +129,20 @@ public class BucketResourceTest {
   @Test
   // create and delete the bucket, checking that the metadata and cloud state match what is expected
   public void createAndDeleteBucketTest() throws Exception {
-    String bucketName = "testbucket_createanddeletebuckettest";
+    String bucketName = createBucketName("createanddeletebuckettest");
     String flightId = "createAndDeleteBucketTest";
 
     // create the bucket and metadata
     GoogleBucketResource bucketResource =
-        createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightId);
+        createBucket(
+            bucketName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightId,
+            null,
+            null,
+            null,
+            true);
 
     // check the bucket and metadata exist
     checkBucketExists(bucketResource.getResourceId());
@@ -138,12 +156,11 @@ public class BucketResourceTest {
   // create buckets in different regions and see if they're actually created there.
   public void createBucketsInDifferentRegionsTest() throws Exception {
     for (GoogleRegion region : List.of(GoogleRegion.US_CENTRAL1, GoogleRegion.US_EAST1)) {
-      String bucketName = "testbucket_bucketregionstest_" + region;
+      String bucketName = createBucketName(region.toString());
       String flightId = "bucketRegionsTest_" + region;
 
       // create the bucket and metadata
-      GoogleBucketResource bucketResource =
-          createBucket(bucketName, projectResource, region, flightId);
+      createBucket(bucketName, projectResource, region, flightId, null, null, null, true);
 
       // Get the Bucket
       Bucket cloudBucket = bucketService.getCloudBucket(bucketName);
@@ -156,12 +173,35 @@ public class BucketResourceTest {
   }
 
   @Test
+  public void testMultiRegionalBucket() throws InterruptedException {
+    String bucketName = createBucketName("multiregiontest");
+    String flightId = "testMultiRegionalBucket";
+
+    // create the bucket and metadata
+    GoogleBucketResource bucketResource =
+        createBucket(
+            bucketName, projectResource, GoogleRegion.US, flightId, null, null, null, true);
+
+    // check the bucket and metadata exist
+    checkBucketExists(bucketResource.getResourceId());
+
+    // autoclass
+    assertTrue("autoclass should be enabled", bucketResource.getAutoclassEnabled());
+    Bucket bucket = storage.get(bucketResource.getName());
+    assertTrue("Autoclass", bucket.getAutoclass().getEnabled());
+
+    // delete the bucket and metadata
+    deleteBucket(bucketResource);
+    checkBucketDeleted(bucketResource.getName(), bucketResource.getResourceId());
+  }
+
+  @Test
   // two threads compete for the bucket lock, confirm one wins and one loses. a third thread fetches
   // the bucket
   // after it's been created, confirm it succeeds.
   public void twoThreadsCompeteForLockTest() throws Exception {
     String flightIdBase = "twoThreadsCompeteForLockTest";
-    String bucketName = "twothreadscompeteforlocktest";
+    String bucketName = createBucketName("twothreadscompetetest");
     GoogleRegion bucketRegion = GoogleRegion.DEFAULT_GOOGLE_REGION;
 
     BucketResourceLockTester resourceLockA =
@@ -235,12 +275,20 @@ public class BucketResourceTest {
         "property allowReuseExistingBuckets = {}",
         bucketResourceUtils.getAllowReuseExistingBuckets(configService));
 
-    String bucketName = "testbucket_bucketexistsbeforemetadatatest";
+    String bucketName = createBucketName("bucketbeforemetadatatest");
     String flightIdA = "bucketExistsBeforeMetadataTestA";
 
     // create the bucket and metadata
     GoogleBucketResource bucketResource =
-        createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightIdA);
+        createBucket(
+            bucketName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightIdA,
+            null,
+            null,
+            null,
+            true);
     checkBucketExists(bucketResource.getResourceId());
 
     // delete the metadata only
@@ -262,7 +310,15 @@ public class BucketResourceTest {
     String flightIdB = "bucketExistsBeforeMetadataTestB";
     boolean caughtCorruptMetadataException = false;
     try {
-      createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightIdB);
+      createBucket(
+          bucketName,
+          projectResource,
+          GoogleRegion.DEFAULT_GOOGLE_REGION,
+          flightIdB,
+          null,
+          null,
+          null,
+          true);
     } catch (CorruptMetadataException cmEx) {
       caughtCorruptMetadataException = true;
     }
@@ -273,7 +329,15 @@ public class BucketResourceTest {
     bucketResourceUtils.setAllowReuseExistingBuckets(configService, true);
     String flightIdC = "bucketExistsBeforeMetadataTestC";
     bucketResource =
-        createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightIdC);
+        createBucket(
+            bucketName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightIdC,
+            null,
+            null,
+            null,
+            false);
 
     // check the bucket and metadata exist
     checkBucketExists(bucketResource.getResourceId());
@@ -292,16 +356,29 @@ public class BucketResourceTest {
         "property allowReuseExistingBuckets = {}",
         bucketResourceUtils.getAllowReuseExistingBuckets(configService));
 
-    String bucketName = "testbucket_nobucketbutmetadataexiststest";
+    String bucketName = createBucketName("nobucketbutmetadatatest");
     String flightIdA = "noBucketButMetadataExistsTestA";
 
     // create the bucket and metadata
     GoogleBucketResource bucketResource =
-        createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightIdA);
+        createBucket(
+            bucketName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightIdA,
+            null,
+            null,
+            null,
+            false);
     checkBucketExists(bucketResource.getResourceId());
 
-    // delete the bucket cloud resource only
     Bucket bucket = storage.get(bucketName);
+
+    // autoclass test
+    assertFalse("autoclass should not be enabled", bucketResource.getAutoclassEnabled());
+    assertNull("Autoclass setting on bucket is null", bucket.getAutoclass());
+
+    // delete the bucket cloud resource only
     boolean bucketDeleted = bucket.delete();
     assertTrue("bucket cloud resource deleted", bucketDeleted);
 
@@ -318,7 +395,15 @@ public class BucketResourceTest {
     String flightIdB = "bucketExistsBeforeMetadataTestB";
     caughtCorruptMetadataException = false;
     try {
-      createBucket(bucketName, projectResource, GoogleRegion.DEFAULT_GOOGLE_REGION, flightIdB);
+      createBucket(
+          bucketName,
+          projectResource,
+          GoogleRegion.DEFAULT_GOOGLE_REGION,
+          flightIdB,
+          null,
+          null,
+          null,
+          true);
     } catch (CorruptMetadataException cmEx) {
       caughtCorruptMetadataException = true;
     }
@@ -330,15 +415,118 @@ public class BucketResourceTest {
     checkBucketDeleted(bucketResource.getName(), bucketResource.getResourceId());
   }
 
+  @Test
+  public void createBucketWithTtlTest() throws Exception {
+    String bucketWithTtlName = createBucketName("bucketwithttltest");
+    String flightWithTtlId = "bucketwithttltest";
+
+    String bucketWithoutTtlName = createBucketName("bucketwithoutttltest");
+    String flightWithoutTtlId = "bucketwithoutttltest";
+
+    Integer deleteAge = 1;
+
+    // create the bucket and metadata
+    GoogleBucketResource bucketWithTtlResource =
+        createBucket(
+            bucketWithTtlName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightWithTtlId,
+            Duration.ofDays(1),
+            null,
+            null,
+            true);
+
+    GoogleBucketResource bucketWithoutTtlResource =
+        createBucket(
+            bucketWithoutTtlName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightWithoutTtlId,
+            null,
+            null,
+            null,
+            true);
+
+    // check the bucket and metadata exist
+    Bucket bucketWithTtl = storage.get(bucketWithTtlName);
+    Bucket bucketWithoutTtl = storage.get(bucketWithoutTtlName);
+
+    var lifecycleRule = bucketWithTtl.getLifecycleRules().get(0);
+    var lifecycleAction = lifecycleRule.getAction();
+    var lifecycleDeleteAge = lifecycleRule.getCondition().getAge();
+
+    assertThat(
+        "Delete lifecycle action was set on the bucket",
+        lifecycleAction.getActionType(),
+        equalTo(BucketInfo.LifecycleRule.LifecycleAction.newDeleteAction().getActionType()));
+
+    assertThat("Delete lifecycle action is set for 1 day", lifecycleDeleteAge, equalTo(deleteAge));
+
+    assertTrue(
+        "Bucket without ttl has no lifecycle rules",
+        bucketWithoutTtl.getLifecycleRules().isEmpty());
+
+    // delete the bucket and metadata
+    for (var bucketResource : List.of(bucketWithTtlResource, bucketWithoutTtlResource)) {
+      deleteBucket(bucketResource);
+      checkBucketDeleted(bucketResource.getName(), bucketResource.getResourceId());
+    }
+  }
+
+  @Test
+  public void testCreateBucketWithReaders() throws Exception {
+    String bucketName = createBucketName("bucketwithreadergroupstest");
+    String flightId = "bucketWithReaderGroupsTest";
+    List<String> readerGroups =
+        new ArrayList<>(
+            samService.retrievePolicyEmails(null, IamResourceType.DATASET, datasetId).values());
+
+    GoogleBucketResource bucket =
+        createBucket(
+            bucketName,
+            projectResource,
+            GoogleRegion.DEFAULT_GOOGLE_REGION,
+            flightId,
+            null,
+            () -> readerGroups,
+            null,
+            true);
+
+    Policy iamPolicy =
+        storage.getIamPolicy(
+            bucket.getName(), Storage.BucketSourceOption.requestedPolicyVersion(3));
+
+    assertThat(
+        "Bucket has dataset users as bucket readers",
+        iamPolicy.getBindingsList(),
+        hasItem(
+            Binding.newBuilder()
+                .setRole(GoogleBucketService.STORAGE_OBJECT_VIEWER_ROLE)
+                .setMembers(List.of("group:jadeteam@broadinstitute.org"))
+                .build()));
+  }
+
   private GoogleBucketResource createBucket(
       String bucketName,
       GoogleProjectResource projectResource,
       GoogleRegion bucketRegion,
-      String flightId)
+      String flightId,
+      Duration ttl,
+      Callable<List<String>> getReaderGroups,
+      String dedicatedServiceAccount,
+      boolean autoclassEnabled)
       throws InterruptedException {
-
     GoogleBucketResource bucketResource =
-        bucketService.getOrCreateBucket(bucketName, projectResource, bucketRegion, flightId);
+        bucketService.getOrCreateBucket(
+            bucketName,
+            projectResource,
+            bucketRegion,
+            flightId,
+            ttl,
+            getReaderGroups,
+            dedicatedServiceAccount,
+            autoclassEnabled);
 
     bucketResources.add(bucketResource);
     datasetBucketDao.createDatasetBucketLink(datasetId, bucketResource.getResourceId());
@@ -358,10 +546,12 @@ public class BucketResourceTest {
   }
 
   private void deleteBucket(GoogleBucketResource bucketResource) {
-    // delete the bucket and update the metadata
-    storage.delete(bucketResource.getName());
-    bucketService.updateBucketMetadata(bucketResource.getName(), null);
-    datasetBucketDao.deleteDatasetBucketLink(datasetId, bucketResource.getResourceId());
+    if (bucketResource != null) {
+      // delete the bucket and update the metadata
+      storage.delete(bucketResource.getName());
+      bucketService.updateBucketMetadata(bucketResource.getName(), null);
+      datasetBucketDao.deleteDatasetBucketLink(datasetId, bucketResource.getResourceId());
+    }
   }
 
   private void checkBucketDeleted(String bucketName, UUID bucketResourceId) {
@@ -388,7 +578,7 @@ public class BucketResourceTest {
     Map<String, List<String>> roleToStewardMap = new HashMap<>();
     roleToStewardMap.put(role, stewardsGroupEmailList);
 
-    ResourceInfo resourceInfo = bufferService.handoutResource();
+    ResourceInfo resourceInfo = bufferService.handoutResource(false);
 
     // create project metadata
     return projectService.initializeGoogleProject(
@@ -396,6 +586,22 @@ public class BucketResourceTest {
         profile,
         roleToStewardMap,
         GoogleRegion.DEFAULT_GOOGLE_REGION,
-        Map.of("test-name", "bucket-resource-test"));
+        Map.of("test-name", "bucket-resource-test"),
+        CollectionType.DATASET);
+  }
+
+  /**
+   * @param baseName the desired prefix for a GCS bucket name
+   * @return a globally-unique GCS bucket name based off of the provided baseName
+   */
+  private String createBucketName(String baseName) {
+    String globallyUniqueName = Names.randomizeName(baseName.toLowerCase());
+    // For more information on GCS naming requirements:
+    // https://cloud.google.com/storage/docs/buckets#naming
+    assertThat(
+        "Bucket name follows GCS naming requirements",
+        globallyUniqueName,
+        matchesPattern("^[a-z0-9][a-z0-9\\-_]{1,61}[a-z0-9]$"));
+    return globallyUniqueName;
   }
 }
